pass

+++ src/main/java/bg/example/recepeWebsite/repository/RecipeRepository.java
@@ -23,6 +23,6 @@ public interface RecipeRepository extends JpaRepository<RecipeEntity, Long>,
     @Query("SELECT r FROM RecipeEntity r JOIN r.favoriteUsers u WHERE u.id = :userId")
     Page<RecipeEntity> findAllFavoriteRecipes(@Param("userId") Long userId, Pageable pageable);
 
-    int countRecipeEntitiesByCategory(CategoryNameEnum category);
+    long countRecipeEntitiesByCategory(CategoryNameEnum category);
 }
